Fixes to nrrd. Affine matrix currently not working, but files load
