+++ test_script.py
@@ -1,6 +1,9 @@
 from glue_medical import setup
 from glue_medical.nifti.nifti_factory import nifti_reader
 
+from glue_medical.nrrd.nrrd_factory import nrrd_reader
+
+
 setup()
 
-nifti_reader('MRBrainTumor2.nii.gz')
+nrrd_reader('MRBrainTumor2.nrrd')
